requirements.txt
biopython==1.76
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
google-auth==1.14.1
idna==2.9
kubernetes==11.0.0
numpy==1.18.3
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
urllib3==1.25.9
websocket-client==0.57.0
